[IMP] sale: keep only onchange tests in test_onchange file
